sync - cl test 03 train model - working well with the optimizations implemented - after copying base_unet weights to new conv_in, panorama quality improved greatly with relevant contents and seams started being smoothy and flip reduced, 20 ep 2 bs 8L4, 13020 data samples

+++ cl/inference/pipeline.py
@@ -23,7 +23,8 @@ class CubeDiffPipeline:
         # Load base SD pipeline
         self.pipeline = StableDiffusionPipeline.from_pretrained(
             pretrained_model_name,
-            torch_dtype=torch.float16 if device == "cuda" else torch.float32,
+            # torch_dtype=torch.float16 if device == "cuda" else torch.float32,
+            torch_dtype=torch.bfloat16 if device == "cuda" else torch.float32,
         )
         self.pipeline.to(device)
